backend/requirements.txt
alembic
sqlalchemy
pyodbc
pymssql
fastapi
uvicorn
pydantic
pyjwt
pathlib
python-jose[cryptography]
passlib[bcrypt]